reorganize and add design document
