+++ src/edu/colostate/cs/cs414/ctrlaltdefeat/System/GymSystemCreator.java
@@ -1,4 +1,4 @@
-package edu.colostate.cs.cs414.ctrlaltdefeat.Controllers;
+package edu.colostate.cs.cs414.ctrlaltdefeat.System;
 
 import java.io.File;
 import java.util.ArrayList;
